refactor: Rename files
camelCase -> kebab-case
