+++ src/key-tree-factory.js
