+++ src/util/count-down-latch.js
@@ -1,0 +1,9 @@
+
+export default initCount => callback => {
+    let count = initCount;
+    return {
+        countDown: (...args) => {
+            if (--count === 0) callback(...args);
+        },
+    };
+};
